+++ functions/theme-setup.php
@@ -18,6 +18,7 @@ function theme_setup() {
 	add_theme_support( 'custom-logo' );
 	add_theme_support( 'unlink-homepage-logo' );
 	add_theme_support( 'block-template-parts' );
+	add_theme_support( 'post-thumbnails' );
 	remove_theme_support( 'core-block-patterns' );
 }
 add_action( 'after_setup_theme', __NAMESPACE__ . '\\theme_setup', 15 );